+++ src/gatsby/node/createPages.js
@@ -10,9 +10,9 @@ const createPaginatedPages = require('gatsby-paginate');
 
 const templatesDirectory = path.resolve(__dirname, '../../templates');
 const templates = {
-  articles: path.resolve(templatesDirectory, 'articles.template.tsx'),
-  article: path.resolve(templatesDirectory, 'article.template.tsx'),
-  author: path.resolve(templatesDirectory, 'author.template.tsx'),
+  articles: path.resolve(templatesDirectory, 'ArticlesTemplate.tsx'),
+  article: path.resolve(templatesDirectory, 'ArticleTemplate.tsx'),
+  author: path.resolve(templatesDirectory, 'AuthorTemplate.tsx'),
 };
 
 const query = require('../data/data.query');
