+++ src/sections/author/AuthorArticles.tsx
@@ -6,7 +6,7 @@ import styled from '@emotion/styled';
 import mediaqueries from '@styles/media';
 import { IArticle } from '@types';
 
-import ArticlesList from '@sections/articles/Articles.List';
+import ArticlesList from '@sections/articles/ArticlesList';
 
 interface AuthorArticlesProps {
   articles: IArticle[];
