+++ src/sections/author/AuthorHero.tsx
